refactor(ktable): emit update:table-preferences
