+++ src/types/index.ts
@@ -3,3 +3,4 @@ export * from './alert'
 export * from './badge'
 export * from './breadcrumbs'
 export * from './button'
+export * from './table'
