+++ src/types/table.ts
@@ -1,0 +1,18 @@
+export interface TablePreferences {
+  /** The number of items to display per page */
+  pageSize?: number
+  /** The column to sort by's `key` defined in the table headers */
+  sortColumnKey?: string
+  /** The order by which to sort the column, one of `asc` or `desc` */
+  sortColumnOrder?: 'asc' | 'desc'
+}
+
+export type TablePaginationType = 'default' | 'offset'
+
+export interface TableHeader {
+  key: string
+  label: string
+  sortable?: boolean
+  hideLabel?: boolean
+  useSortHandlerFn?: boolean
+}
